requirements/base.txt
Flask==1.1.1
psycopg2-binary==2.8.4
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
python-dotenv==0.12.0